% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{binProbs}
\alias{binProbs}
\title{Get binned probabilities.}
\usage{
binProbs(y.true, y.prob, nbins = 5)
}
\arguments{
\item{y.true}{A numeric vector. The true response.}

\item{y.prob}{A numeric vector. The predicted probabilites.}

\item{nbins}{An integer. The number of bins.}
}
\value{
A list with two elements. The first element is the binned actual fractions, and
the second element is the binned predicted probabilites.
}
\description{
Get binned probabilities.
}
\examples{
y.true <- c(0, 0, 1, 1, 0, 1, 1, 0, 0, 0)
y.prob <- c(0.45454545, 0.36363636, 0.63636364, 0.18181818, 0.45454545, 0.09090909,
0.27272727, 0.81818182, 0.63636364, 0.63636364)
binProbs(y.true, y.prob)

}
