% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{brier}
\alias{brier}
\title{Standard Brier Score}
\usage{
brier(y.true, y.prob)
}
\arguments{
\item{y.true}{A numeric vector. The true response.}

\item{y.prob}{A numeric vector. The predicted probabilites.}
}
\value{
A numeric scalar. The standard Brier score.
}
\description{
Standard Brier Score
}
\details{
Brier score measures the fit of probability estimates to the observed data.
It is defined as the mean squared difference between the observed labels the
estimated probability. A smaller value means a better calibration.
\deqn{BS = \frac{\sum_{i=1}^N (y_i - \hat{P}(y_i | x_i))^2}{N}}
}
\examples{
y.true <- c(0, 0, 1, 1, 0, 1, 1, 0, 0, 0)
y.prob <- c(0.45454545, 0.36363636, 0.63636364, 0.18181818, 0.45454545, 0.09090909,
0.27272727, 0.81818182, 0.63636364, 0.63636364)
brier(y.true, y.prob)

}
