% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{comparisonPlot}
\alias{comparisonPlot}
\title{Plot comparison among calibrations.}
\usage{
comparisonPlot(y.true, probs, mod.names, nbins = 5)
}
\arguments{
\item{y.true}{A numeric vector. The true response.}

\item{nbins}{An integer. The number of bins.}

\item{y.prob}{A numeric vector. The predicted probabilites.}

\item{mod.name}{A string. The name of the model you want to plot, used in the legend.}
}
\value{
A plot with calibration curves given by different calibrations.
}
\description{
Do comparisons among different models or different calibration methods, in terms of the
calibrated probabilities.
}
\examples{
library('e1071')
data(imbalance)
set.seed(123)
split <- sample.split(imbalance$y, SplitRatio = 0.75)
train_set <- subset(imbalance, split == TRUE)
test_set <- subset(imbalance, split == FALSE)
X.test <- subset(test_set, select = -y)
y.test <- subset(test_set, select = y)[,1]
lr <- glm(y ~ ., data = train_set, family = "binomial")
prob.lr <- predict(lr, X.test, type = "response")
calibCurve(y.test, prob.lr, "Logistic")
nb <- naiveBayes(y ~ ., data = train_set)
prob.nb <- predict(nb, X.test, type = "raw")[, 2]
calibCurve(y.test, prob.nb, "Naive Bayes")
comparisonPlot(y.test, list(prob.lr, prob.nb), c("Logistic", "Naive Bayes"))

}
\references{
\href{https://scikit-learn.org/stable/auto_examples/calibration/plot_compare_calibration.html}{sklearn: Comparison of Calibration of Classifiers}
}
