% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{stratifiedBrier}
\alias{stratifiedBrier}
\title{Stratified Brier Score}
\usage{
stratifiedBrier(y.true, y.prob)
}
\arguments{
\item{y.true}{A numeric vector. The true response.}

\item{y.prob}{A numeric vector. The predicted probabilites.}
}
\value{
A list with three elements. The 1st is the standard Brier score, The 2nd is the Brier score
for the positive class, and the 3rd is the Brier score for the negative class.
}
\description{
Stratified Brier Score
}
\details{
Stratified Brier Score evaluates the goodness of calibration under the imbalanced scenario.
\deqn{BS^+ = \frac{\sum_{y_i=\text{pos_label}} (y_i - \hat{P}(y_i | x_i))^2}{N_{pos}}}
\deqn{BS^- = \frac{\sum_{y_i=\text{neg_label}} (y_i - \hat{P}(y_i | x_i))^2}{N_{pos}}}
}
\examples{
y.true <- c(0, 0, 1, 1, 0, 1, 1, 0, 0, 0)
y.prob <- c(0.45454545, 0.36363636, 0.63636364, 0.18181818, 0.45454545, 0.09090909,
0.27272727, 0.81818182, 0.63636364, 0.63636364)
stratifiedBrier(y.true, y.prob)

}
\references{
Wallace, B.C., Dahabreh, I.J. Improving class probability estimates for imbalanced data.
Knowl Inf Syst 41, 33–52 (2014). https://doi.org/10.1007/s10115-013-0670-6
}
