% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate.R
\name{underSample}
\alias{underSample}
\title{Dataset under-sampling}
\usage{
underSample(pos_set, neg_set)
}
\arguments{
\item{pos_set}{A dataframe. The dataframe for the positive class.}

\item{neg_set}{A dataframe. The dataframe for the negative class.}
}
\value{
A dataframe. An under-sampled dataset, with equal number of positive class and
negative class.
}
\description{
Dataset under-sampling
}
